refactor: ♻️ Refactor AppleNotificationController, improve test coverage, and update workflow.
