+++ tests/Browser/console/GeneaLabs_LaravelSignInWithApple_Tests_Browser_SignInWithAppleTest_testRedirectShowsAppleIdLogin-0.log
@@ -1,8 +1,20 @@
 [
     {
         "level": "WARNING",
-        "message": "https:\/\/appleid.cdn-apple.com\/appleauth\/static\/jsj\/2086116106\/profile\/app.js 254 Mixed Content: The page at 'https:\/\/appleid.apple.com\/auth\/authorize?client_id=dev.test.service-id&redirect_uri=http%3A%2F%2Ftesting.dev%2Fsiwa-callback&scope=name%20email&response_type=code&response_mode=form_post&state=DCCb1UiYHkFyhepDPSMCvHorrEQPIfQL4vD7dMZl' was loaded over a secure connection, but contains a form that targets an insecure endpoint 'http:\/\/testing.dev\/siwa-callback'. This endpoint should be made available over a secure connection.",
-        "source": "security",
-        "timestamp": 1583082984883
+        "message": "http:\/\/127.0.0.1:8001\/tests - Failed to decode downloaded font: data:application\/x-font-woff;charset=utf-8;base64,d09GRgABAAAAABRMABEAAAAAIawAAQAAAAAAAAAAAAAAAAAAAAAAAAAAAABHUE9TAAATFAAAALsAAAHIbUB2PEdTVUIAABPQAAAAZQAAAIxKSyvpT1MvMgAACjgAAABNAAAAYHLeeipic2xuAAAUOAAAABMAAABI\/ykCnmNtYXAAAAqIAAAArAAAATzUgYTCY3Z0IAAAEagAAACGAAAA\/h4jsglmcGdtAAALNAAABcMAAAviP64gqWdhc3AAABMMAAAACAAAAAgAAAAQZ2x5ZgAAAYAAAAfMAAAMDN+ERypoZWFkAAAJlAAAADYAAAA2FZUeyWhoZWEAAAoYAAAAIAAAACQQagbMaG10eAAACcwAAABMAAAATFWqCFBsb2NhAAAJbAAAACgAAAAoG5oe821heHAAAAlMAAAAIAAAACABaQyMbmFtZQAAEjAAAADFAAABhhtRNi1wb...0IQ2XzXnn7yuvSUW0L\/9kXpBkTdbF+9L37sSPb8Jyvv8\/fASPuNJwAAAB4nGNgZgCDfzcYZjFgAQA4VwJ0AAABAAH\/\/wAPeJyNUDEOwjAMPCcF2lQChPoAHsDICxBiYmRkQYiJqkPFAH9jZGIF8RIWxGAuKVI7dGhOsZ3zObINAeCwwBZ2uVpvkO0vZY5pvjsVmCFiFqrwqjqWRmxgj4eyQNa0TEq4EZLg46AEvRHjq2Uic6QE9Ko34q5ntB59tfMtyo8+O2sfXZW+A\/b3bbzf1fzdav++ns4E+L2kGIfNWAyrTfLfHvoE6AdETJ0LuRFZIeNrPZvQOsauqvoB5z0tQgB4nGNgZGBg4GKIYihhYHZx8wlhEEmuLMphkMtJLMljUGJgAcoy\/P\/PAAPMjlGuCgxizkEhCgxyIUHeCgxqYHlGqDpGEAtMMzEw5+Qn5zCIIJNARYxgzAKlOYCYDawLyAYAAaIWnAAAAHicY2BkgAKmef81GMgGAGutAckA",
+        "source": "other",
+        "timestamp": 1774830557340
+    },
+    {
+        "level": "WARNING",
+        "message": "http:\/\/127.0.0.1:8001\/tests - OTS parsing error: name: Failed to parse table",
+        "source": "other",
+        "timestamp": 1774830557340
+    },
+    {
+        "level": "SEVERE",
+        "message": "https:\/\/symflower.com\/siwa-callback - Failed to load resource: the server responded with a status of 404 ()",
+        "source": "network",
+        "timestamp": 1774830557739
     }
 ]
