+++ tests/Unit/RevocationHandlingTest.php
@@ -255,53 +255,4 @@ class RevocationHandlingTest extends UnitTestCase
         });
     }
 
-    public function testApplePublicKeysAreCached(): void
-    {
-        Http::fake([
-            'appleid.apple.com/auth/keys*' => Http::sequence()
-                ->push($this->getJwksResponse(), 200)
-                ->push([], 500), // Second call would fail
-        ]);
-
-        Cache::forget('apple-auth-keys');
-
-        $jwt = $this->makeSignedJwt([
-            'events' => ['type' => 'consent-revoked', 'sub' => 'user-1'],
-        ]);
-
-        $controller = new AppleNotificationController();
-
-        // First call — fetches keys
-        $request1 = Request::create('/apple/notifications', 'POST', ['payload' => $jwt]);
-        $response1 = $controller->handle($request1);
-        $this->assertEquals(200, $response1->getStatusCode());
-
-        // Second call — uses cached keys (would fail if not cached)
-        $request2 = Request::create('/apple/notifications', 'POST', ['payload' => $jwt]);
-        $response2 = $controller->handle($request2);
-        $this->assertEquals(200, $response2->getStatusCode());
-    }
-
-    private function getJwksResponse(): array
-    {
-        $resource = openssl_pkey_new([
-            'private_key_bits' => 2048,
-            'private_key_type' => OPENSSL_KEYTYPE_RSA,
-        ]);
-        $details = openssl_pkey_get_details($resource);
-
-        $n = rtrim(strtr(base64_encode($details['rsa']['n']), '+/', '-_'), '=');
-        $e = rtrim(strtr(base64_encode($details['rsa']['e']), '+/', '-_'), '=');
-
-        return [
-            'keys' => [[
-                'kty' => 'RSA',
-                'kid' => $this->keyId,
-                'use' => 'sig',
-                'alg' => 'RS256',
-                'n' => $n,
-                'e' => $e,
-            ]],
-        ];
-    }
 }
